refactor(specification): create module for specification
