+++ src/modules/specification/models/Specification.ts
