+++ src/modules/specification/repositories/ISpecificationRepository.ts
